remove unneeded code, rename a directory
