+++ luaAITB/callbacks/evaluateCache.lua
@@ -1,4 +1,4 @@
-local sheepStats = require("luaAITB.ampharos.sheepStats")
+local sheepStats = require("luaAITB.characters.sheepStats")
 
 modAITB:AddCallback(ModCallbacks.MC_EVALUATE_CACHE, function(_, player, cacheFlag)
     sheepStats:calcStats(player, cacheFlag)
