+++ luaAITB/callbacks/postPlayerInit.lua
@@ -1,4 +1,4 @@
-local sheepCosts = require("luaAITB.ampharos.sheepCosts")
+local sheepCosts = require("luaAITB.characters.sheepCosts")
 
 modAITB:AddCallback(ModCallbacks.MC_POST_PLAYER_INIT, function(_, player)
     sheepCosts:applyCosts(player)
